pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apa-formatter"
version = "0.1.0"
description = "Formateador de documentos Word y PDF conforme a normas APA 7ª edición"
readme = "README.md"
license = "GPL-3.0"
requires-python = ">=3.10"
authors = [{ name = "mackroph" }]

dependencies = [
    "python-docx>=1.1.0",
    "fpdf2>=2.8.0",
    "typer[all]>=0.12.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "bibtexparser>=1.4.0",
    "nameparser>=1.1.0",
    "platformdirs>=4.0.0",
    "pdfplumber>=0.10.0",
]

[project.optional-dependencies]
gui = [
    "PySide6>=6.6.0",
]
ai = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.0.0",
]

[project.scripts]
apa = "apa_formatter.presentation.cli.app:app"
apa-gui = "apa_formatter.gui.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/apa_formatter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = "apa_formatter.gui.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "apa_formatter.presentation.gui.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "requests", "requests.*",
    "bs4", "bs4.*",
    "apa_formatter.fetchers.*",
]
ignore_missing_imports = true
disallow_untyped_defs = false
